+++ packages/usage/lib/index.ts
@@ -1,30 +1,36 @@
 import { getKVStore } from '@nangohq/kvstore';
 
-import { DbUsageStore } from './usageStore/dbUsageStore.js';
-import { HybridUsageStore } from './usageStore/hybridUsageStore.js';
-import { KvUsageStore } from './usageStore/kvUsageStore.js';
-import { UsageTracker } from './usageTracker.js';
+import { DbAccountUsageStore } from './accountUsageStore/dbAccountUsageStore.js';
+import { HybridAccountUsageStore } from './accountUsageStore/hybridAccountUsageStore.js';
+import { KvAccountUsageStore } from './accountUsageStore/kvAccountUsageStore.js';
+import { AccountUsageTracker } from './accountUsageTracker.js';
 
-export { UsageTracker } from './usageTracker.js';
-export type { UsageStore } from './usageStore/usageStore.js';
-export { KvUsageStore } from './usageStore/kvUsageStore.js';
-export { HybridUsageStore } from './usageStore/hybridUsageStore.js';
+export { AccountUsageTracker } from './accountUsageTracker.js';
+export type { AccountUsageStore } from './accountUsageStore/accountUsageStore.js';
+export { DbAccountUsageStore } from './accountUsageStore/dbAccountUsageStore.js';
+export { HybridAccountUsageStore } from './accountUsageStore/hybridAccountUsageStore.js';
+export { KvAccountUsageStore } from './accountUsageStore/kvAccountUsageStore.js';
 
-let usageTracker: UsageTracker | undefined;
+let usageTracker: AccountUsageTracker | undefined;
 
-async function createUsageTracker(): Promise<UsageTracker> {
+/**
+ * Creates an AccountUsageTracker with a HybridAccountUsageStore.
+ * Focused on performance over precision.
+ * Intended for simple local account usage tracking, not for billing.
+ */
+async function createAccountUsageTracker(): Promise<AccountUsageTracker> {
     const kvStore = await getKVStore();
-    const kvUsageStore = new KvUsageStore(kvStore);
-    const persistentUsageStore = new DbUsageStore();
-    const usageStore = new HybridUsageStore(kvUsageStore, persistentUsageStore);
-    return new UsageTracker(usageStore);
+    const kvUsageStore = new KvAccountUsageStore(kvStore);
+    const persistentUsageStore = new DbAccountUsageStore();
+    const usageStore = new HybridAccountUsageStore(kvUsageStore, persistentUsageStore);
+    return new AccountUsageTracker(usageStore);
 }
 
-export async function getUsageTracker(): Promise<UsageTracker> {
+export async function getAccountUsageTracker(): Promise<AccountUsageTracker> {
     if (usageTracker) {
         return usageTracker;
     }
 
-    usageTracker = await createUsageTracker();
+    usageTracker = await createAccountUsageTracker();
     return usageTracker;
 }
